requirements.txt
requests>=2.20.0
websocket-client>=1.0.0
